feat: Seek to event when clicking it
